Move AbstractSpec
