+++ shake/shakespeare/nodes/src/commonMain/kotlin/com/shakelang/shake/shakespeare/nodes/spec/AbstractNodeSpec.kt
@@ -1,7 +1,7 @@
 package com.shakelang.shake.shakespeare.nodes.spec
 
 import com.shakelang.shake.parser.node.ShakeNode
-import com.shakelang.shake.shakespeare.AbstractSpec
+import com.shakelang.shake.shakespeare.spec.AbstractSpec
 import com.shakelang.shake.shakespeare.spec.GenerationContext
 
 interface AbstractNodeSpec : AbstractSpec {
