--- shake/shakespeare/src/commonMain/kotlin/com/shakelang/shake/shakespeare/AbstractSpec.kt
@@ -1,7 +1,0 @@
-package com.shakelang.shake.shakespeare
-
-import com.shakelang.shake.shakespeare.spec.GenerationContext
-
-interface AbstractSpec {
-    fun generate(ctx: GenerationContext): String
-}
