+++ shake/shakespeare/src/commonMain/kotlin/com/shakelang/shake/shakespeare/spec/AbstractSpec.kt
@@ -1,0 +1,5 @@
+package com.shakelang.shake.shakespeare.spec
+
+interface AbstractSpec {
+    fun generate(ctx: GenerationContext): String
+}
