+++ shake/shakespeare/src/commonMain/kotlin/com/shakelang/shake/shakespeare/spec/code/Statements.kt
@@ -8,7 +8,7 @@
 
 package com.shakelang.shake.shakespeare.spec.code
 
-import com.shakelang.shake.shakespeare.AbstractSpec
+import com.shakelang.shake.shakespeare.spec.AbstractSpec
 import com.shakelang.shake.shakespeare.spec.GenerationContext
 import com.shakelang.shake.shakespeare.spec.TypeSpec
